+++ src/prelude.rs
@@ -2,4 +2,5 @@ pub use crate::StripFx;
 pub use crate::effect::Effect;
 pub use crate::segment::{EffectConfig, EffectState, Segment, SegmentOptions};
 pub use crate::utils::{BLACK, BLUE, GREEN, ORANGE, PURPLE, RED, WHITE, rgb};
+pub use color8::CrgbPalette16;
 pub use smart_leds_trait::RGB8;